Reorganize swss-common types
